% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{alleleDiff_rna}
\alias{alleleDiff_rna}
\title{Run DESeq2 with interaction design}
\usage{
alleleDiff_rna(rnaCountObject, fdrCutoff = 0.01, tfname = "mof")
}
\arguments{
\item{rnaCountObject}{output from countFeatures_rna script}

\item{fdrCutoff}{FDR cutoff for differential expression}

\item{tfname}{Name of TF to get the differential expression (must match the name in samplesheet)}
}
\value{
rnaResultObject
}
\description{
Run DESeq2 with interaction design
}
\examples{
alleleDiff_rna(rnaCountObject,fdrCutoff = 0.01,tfname = "mof")

}

