% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{countFeatures_rna}
\alias{countFeatures_rna}
\title{Read the Files and Count features for the RNA-Seq Samples}
\usage{
countFeatures_rna(csvFile = "testBAMs/testSampleSheet.csv", AnnotationFile,
  nthreads, outfileName = NULL, refAllele = "pat", userParam = FALSE, ...)
}
\arguments{
\item{csvFile}{csvfile with sample information}

\item{AnnotationFile}{GTF file}

\item{nthreads}{Number of threads}

\item{outfileName}{Specify output file}

\item{refAllele}{Reference Allele (name must match the name in samplesheet)}

\item{userParam}{Make it TRUE if you want to enter your own featurecounts parameters}
}
\value{
rnaCountObject : a list with counts and sampleinfo
}
\description{
Read the Files and Count features for the RNA-Seq Samples
}
\examples{
countFeatures_rna(csvFile = "testBAMs/testSampleSheet.csv",AnnotationFile,nthreads,
outfileName=NULL,refAllele = "pat",userParam = FALSE,...)

}

