% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASB_functions.R
\name{makeQCplots_chip}
\alias{makeQCplots_chip}
\title{Make plots to select window size and pe-distance cutoffs}
\usage{
makeQCplots_chip(chipCountObject, outdir)
}
\arguments{
\item{chipCountObject}{output from readfiles_chip}

\item{outdir}{output directory to write the plots}
}
\value{
chipCountObject : a list with window counts and sampleinfo
}
\description{
Make plots to select window size and pe-distance cutoffs
}
\examples{
makeQCplots_chip(chipCountObject,outdir)

}

