% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{plotResults_rna}
\alias{plotResults_rna}
\title{Make QC and result plots for DESeq2 output}
\usage{
plotResults_rna(rnaResultObject, outfile = "resultPlots_RNA.pdf",
  barplot = TRUE, excludeChr = "chr12")
}
\arguments{
\item{rnaResultObject}{Output of alleleDiff_rna}

\item{outfile}{Output file to write back the result}

\item{barplot}{TRUE/FALSE whether you want the barplot for allele-biased gene numbers}

\item{excludeChr}{The chromosome to exclude from the barplot (if barplot = TRUE)}
}
\value{
A pdf file with QC and results plots
}
\description{
Make QC and result plots for DESeq2 output
}
\examples{
plotResults_rna(rnaResultObject,outfile = "resultPlots_RNA.pdf")
}

