% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cSaw_functions.R
\name{readfiles_chip}
\alias{readfiles_chip}
\title{Read the Files and Count windows for ChIP-Seq Samples}
\usage{
readfiles_chip(csvFile = "testBAMs/testSampleSheet.csv", refAllele = "pat")
}
\arguments{
\item{csvFile}{csvfile with sample information}

\item{refAllele}{Reference Allele (name must match the name in samplesheet)}
}
\value{
chipCountObject : a list with window counts and sampleinfo
}
\description{
Read the Files and Count windows for ChIP-Seq Samples
}
\examples{
readfiles_chip(csvFile = "testBAMs/testSampleSheet.csv", refAllele = "pat")
}

