% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{writeOutput_rna}
\alias{writeOutput_rna}
\title{Write the output}
\usage{
writeOutput_rna(rnaResultObject, annotateFrom = "dataset",
  species = "Mus musculus", excludeChr = "chr12", fdrCutoff = 0.01,
  outfileName)
}
\arguments{
\item{rnaResultObject}{output from alleleDiff_rna function}

\item{annotateFrom}{either "dataset","ensembl" or NULL. telling where to annotate the output from}

\item{species}{Species for annotation (if annotateFrom = "ensembl")}

\item{excludeChr}{any chromosome to exclude from output}

\item{fdrCutoff}{fdr cutoff (same as in previous script)}

\item{outfileName}{File to write the output in}
}
\value{
csv file with output
}
\description{
Write the output
}
\examples{
writeOutput_rna(rnaResultObject,annotateFrom = "dataset", species = "Mus musculus",
excludeChr = "chr12",fdrCutoff = 0.01,outfileName)
}

